+++ serial_driver_test/arm_pl011.h
@@ -57,6 +57,6 @@ typedef struct {
 
 
 // functions
-extern void pl011_init(PL011_Type *serial, uint32_t uart_clock);
+extern void pl011_init(PL011_Type *serial, uint64_t uart_clock);
 extern int pl011_putc(int c, PL011_Type *serial);
 extern int pl011_getc(PL011_Type *serial);
